requirements/prod.txt
asgiref==3.6.0
Django==4.1.6
django-model-utils==4.3.1
Pillow==9.4.0
psycopg2==2.9.5
sqlparse==0.4.3
tzdata==2022.7
Unipath==1.1